requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
python-dotenv
httpx